+++ tests/protocols/gmpv208/testcmds/test_create_scanner.py
@@ -20,7 +20,7 @@ import unittest
 
 from gvm.errors import RequiredArgument, InvalidArgumentType
 
-from gvm.protocols.gmpv9 import ScannerType
+from gvm.protocols.gmpv208 import ScannerType
 
 
 class GmpCreateScannerTestCase:
