+++ tests/protocols/gmpv208/testcmds/test_get_assets.py
@@ -20,7 +20,7 @@ import unittest
 
 from gvm.errors import InvalidArgumentType
 
-from gvm.protocols.gmpv7 import AssetType
+from gvm.protocols.gmpv208 import AssetType
 
 
 class GmpGetAssetsTestCase:
