+++ tests/protocols/gmpv208/testcmds/test_modify_note.py
@@ -22,7 +22,7 @@ from decimal import Decimal
 
 from gvm.errors import RequiredArgument, InvalidArgumentType
 
-from gvm.protocols.gmpv7 import SeverityLevel
+from gvm.protocols.gmpv208 import SeverityLevel
 
 
 class GmpModifyNoteTestCase:
