--- tests/protocols/gmpv214/test_v8_from_gmpv214.py
@@ -1,75 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2019-2021 Greenbone Networks GmbH
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-from . import Gmpv214TestCase
-from ..gmpv8.testcmds import *  # pylint: disable=unused-wildcard-import,wildcard-import
-
-
-class Gmpv214CloneTicketTestCase(GmpCloneTicketTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214CreateCredentialTestCase(
-    GmpCreateCredentialTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214CreateScheduleTestCase(GmpCreateScheduleTestCase, Gmpv214TestCase):
-
-    pass
-
-
-class Gmpv214CreateTicketTestCase(GmpCreateTicketTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214DeleteTicketTestCase(GmpDeleteTicketTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214GetTicketTestCase(GmpGetTicketTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214GetTargetsTestCase(GmpGetTargetsTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214GetVulnerabilitiesTestCase(
-    GmpGetVulnerabilitiesTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214GetVulnerabilityTestCase(
-    GmpGetVulnerabilityTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214ModifyTestCase(GmpModifyTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214ModifyScheduleTestCase(GmpModifyScheduleTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214ModifyTicketTestCase(GmpModifyTicketTestCase, Gmpv214TestCase):
-    pass
